Move Identification from Plugins to Core
