+++ Tests/DownstreamProject/CMakeLists.txt
@@ -10,7 +10,6 @@ find_package(
     Fatras
     PluginAutodiff
     PluginDD4hep
-    PluginIdentification
     PluginJson
     PluginLegacy
     PluginTGeo)
@@ -29,7 +28,6 @@ target_link_libraries(
     ActsCore
     ActsFatras
     ActsPluginDD4hep
-    ActsPluginIdentification
     ActsPluginJson
     ActsPluginLegacy
     ActsPluginTGeo)
