+++ docs/Doxyfile
@@ -166,7 +166,7 @@ STRIP_FROM_INC_PATH    = ../Core/include \
                          ../Plugins/DD4hep/include \
                          ../Plugins/Digitization/include \
                          ../Plugins/Geant4/include \
-                         ../Plugins/Identification/include \
+                         ../Identification/include \
                          ../Plugins/Json/include \
                          ../Plugins/Legacy/include \
                          ../Plugins/TGeo/include
@@ -779,7 +779,7 @@ INPUT                  = ../Core/include \
                          ../Plugins/Cuda/include \
                          ../Plugins/DD4hep/include \
                          ../Plugins/Digitization/include \
-                         ../Plugins/Identification/include \
+                         ../Identification/include \
                          ../Plugins/Json/include \
                          ../Plugins/Legacy/include \
                          ../Plugins/Onnx/include \
